Implement Tilt and streamline PR CI cycle Add Playwright UI testing
