--- playwright/utils/data/kind/local-kind-config.yaml
@@ -1,10 +1,0 @@
-kind: Cluster
-apiVersion: kind.x-k8s.io/v1alpha4
-nodes:
-- role: control-plane
-  extraMounts:
-    - hostPath: /var/run/docker.sock
-      containerPath: /var/run/docker.sock
-  extraPortMappings:
-  - containerPort: 30080
-    hostPort: 30080
