--- playwright/utils/scripts/resources/ingress/certificate-issuer.yaml
@@ -1,23 +1,0 @@
-apiVersion: cert-manager.io/v1
-kind: Issuer
-metadata:
-  name: selfsigned-issuer
-  namespace: flux-system
-spec:
-  selfSigned: {}
-
----
-apiVersion: cert-manager.io/v1
-kind: Certificate
-metadata:
-  name: selfsigned-ca
-  namespace: flux-system
-spec:
-  isCA: true
-  secretName: weave-gitops-tls
-  issuerRef:
-    name: selfsigned-issuer
-    kind: Issuer
-    group: cert-manager.io
-  dnsNames: [{{HOST_NAME}}]
-  usages: ["server auth", "client auth", "signing", "digital signature", "key encipherment"]
