--- playwright/utils/scripts/resources/shared-secrets-kustomization.yaml
@@ -1,17 +1,0 @@
----
-apiVersion: kustomize.toolkit.fluxcd.io/v1
-kind: Kustomization
-metadata:
-  name: shared-secrets
-  namespace: flux-system
-spec:
-  interval: 10m0s
-  sourceRef:
-    kind: GitRepository
-    name: flux-system
-  path: ./shared-secrets
-  prune: true
-  decryption:
-    provider: sops
-    secretRef:
-      name: sops-gpg
