+++ tools/dev-resources/cert-manager/kustomization.yaml
@@ -1,3 +1,3 @@
 resources:
 - cert-manager.yaml
-- cluster-issuer.yaml
+
