+++ tools/dev-resources/certs/kustomization.yaml
@@ -1,0 +1,3 @@
+resources:
+- cluster-issuer.yaml
+
