+++ tools/dev-resources/namespace.yaml
@@ -12,4 +12,9 @@ metadata:
 apiVersion: v1
 kind: Namespace
 metadata:
+  name: flux-system
+---
+apiVersion: v1
+kind: Namespace
+metadata:
   name: ingress-nginx
